Refactor ISSN-registry publication list and item views
