+++ src/frontend/views/issn-registry/requests/IssnRequest.jsx
@@ -130,7 +130,8 @@ function IssnRequest() {
                 setIssnRequest={setIssnRequest}
                 isEdit={isEdit}
               />
-            </IssnRequestEditForm>}
+            </IssnRequestEditForm>
+          }
         </div>
       </div>
     </>
